+++ js/util/locale.js
@@ -13,7 +13,7 @@ var appName = 'Taivas TV7';
 /**
  * Updated by gulp script during a build of the app.
  */
-var appVersion = '2.2.5';
+var appVersion = '2.2.6';
 
 var logoTaivas = 'images/logo_taivas.png';
 var logoTaevas = 'images/logo_taevas.png';
@@ -63,6 +63,7 @@ var localeTextFi = [
     { id: 'favoritesIconText', text: 'Suosikit' },
     { id: 'channelInfoIconText', text: 'Taivas TV7' },
     { id: 'platformInfoIconText', text: 'Tietoja' },
+    { id: 'savedSearchTitle', text: 'Haku' },
     { id: 'searchText', text: 'Hae' },
     { id: 'searchResultText', text: 'Haku tulokset' },
     { id: 'clearText', text: 'Tyhjennä' },
@@ -100,6 +101,7 @@ var localeTextEt = [
     { id: 'favoritesIconText', text: 'Lemmikud' },
     { id: 'channelInfoIconText', text: 'Taevas TV7' },
     { id: 'platformInfoIconText', text: 'Infot' },
+    { id: 'savedSearchTitle', text: 'Otsing' },
     { id: 'searchText', text: 'Otsi' },
     { id: 'searchResultText', text: 'Otsingu tulemused' },
     { id: 'clearText', text: 'Kustuta' },
@@ -137,6 +139,7 @@ var localeTextRu = [
     { id: 'favoritesIconText', text: 'Избранные' },
     { id: 'channelInfoIconText', text: 'Небеса ТВ7' },
     { id: 'platformInfoIconText', text: 'Информация' },
+    { id: 'savedSearchTitle', text: 'Поиск' },
     { id: 'searchText', text: 'Найти' },
     { id: 'searchResultText', text: 'Результаты поиска' },
     { id: 'clearText', text: 'стере́ть' },
@@ -174,6 +177,7 @@ var localeTextSv = [
     { id: 'favoritesIconText', text: 'Favoriter' },
     { id: 'channelInfoIconText', text: 'Himlen TV7' },
     { id: 'platformInfoIconText', text: 'Information' },
+    { id: 'savedSearchTitle', text: 'Sök' },
     { id: 'searchText', text: 'Sök' },
     { id: 'searchResultText', text: 'Sökresultat' },
     { id: 'clearText', text: 'Radera' },